Seperated authentication from user
