+++ MyMoney.Web/ClientApp/src/app/shared/api/index.ts
@@ -2,3 +2,4 @@ export * from './transaction.api';
 export * from './budget.api';
 export * from './dtos.interface';
 export * from './income.api';
+export * from './authentication.api';
